Add iterations to the Solver algorithm
Each iteration tries to solve the level by using the last iteration's output as this iteration's input. The output of an iteration is the GameState with the highest score, and this score is level specific.

Note that with this commit, the algorithm can actually solve the Floatiest Platforms level of Baba Is You. Hooray!