Take into account turn depth when caching game states
When there's a cache hit, keep the game state from higher in the move tree. Equivalent game states that are lower in the tree should be culled from the tree.

Note that, as a corollary, the hash and equality functions for GameState no longer take into account the turn number.